Convert pathapp to composition api, add usestate to useImageQueue
