+++ client/pages/pathapp/index.vue
@@ -1,0 +1,3 @@
+<template>
+    <div>This is the task picker.</div>
+</template>
